refactor: basic_types.py
Removed __unwrap_annotated, leaving only SerializeAs left,
so moved the file to hint_types\serialize_as.py

Removed unused specialized from utils
