+++ structured/__init__.py
@@ -2,7 +2,6 @@ __author__ = 'lojack5'
 __version__ = '3.0.0'
 
 from .base_types import *
-from .basic_types import *
 from .hint_types import *
 from .serializers import *
 from .structured import *
