+++ structured/hint_types/__init__.py
@@ -1,3 +1,4 @@
 from .arrays import *
 from .basic_types import *
+from .serialize_as import *
 from .strings import *
