+++ structured/serializers/structs.py
@@ -6,9 +6,9 @@ __all__ = [
     'noop_action',
 ]
 
-import sys
 import re
 import struct
+import sys
 from functools import cached_property, partial, reduce
 from itertools import chain, repeat
 from typing import overload
@@ -28,7 +28,6 @@ from ..type_checking import (
     Unpack,
 )
 from .api import Serializer
-
 
 _PY_3_12 = sys.version_info >= (3, 12)
 
@@ -79,6 +78,7 @@ class StructSerializer(Generic[Unpack[Ts]], struct.Struct, Serializer[Unpack[Ts]
     """A Serializer that is a thin wrapper around struct.Struct, class creation
     is cached.
     """
+
     num_values: int
 
     @property
@@ -93,13 +93,20 @@ class StructSerializer(Generic[Unpack[Ts]], struct.Struct, Serializer[Unpack[Ts]
     def _split_format(self) -> tuple[ByteOrder, str]:
         return split_byte_order(self.format)
 
-    def __new__(cls, format: str, num_values: int = 1, byte_order: ByteOrder = ByteOrder.DEFAULT) -> Self:
+    def __new__(
+        cls, format: str, num_values: int = 1, byte_order: ByteOrder = ByteOrder.DEFAULT
+    ) -> Self:
         if _PY_3_12:
-            return super().__new__(cls, byte_order.value + format) # type: ignore
+            return super().__new__(cls, byte_order.value + format)  # type: ignore
         else:
             return super().__new__(cls)
 
-    def __init__(self, format: str, num_values: int = 1, byte_order: ByteOrder = ByteOrder.DEFAULT) -> None:
+    def __init__(
+        self,
+        format: str,
+        num_values: int = 1,
+        byte_order: ByteOrder = ByteOrder.DEFAULT,
+    ) -> None:
         if not _PY_3_12:
             super().__init__(byte_order.value + format)
         self.num_values = num_values
@@ -214,19 +221,31 @@ class StructActionSerializer(Generic[Unpack[Ts]], StructSerializer[Unpack[Ts]]):
     """A Serializer acting as a thin wrapper around struct.Struct, with
     transformations applied to unpacked values.
     """
+
     actions: tuple[Callable[[Any], Any], ...]
 
-    def __new__(cls, fmt: str, num_attrs: int = 1, byte_order: ByteOrder = ByteOrder.DEFAULT, actions: tuple[Callable[[Any], Any], ...] = ()) -> Self:
+    def __new__(
+        cls,
+        fmt: str,
+        num_attrs: int = 1,
+        byte_order: ByteOrder = ByteOrder.DEFAULT,
+        actions: tuple[Callable[[Any], Any], ...] = (),
+    ) -> Self:
         return super().__new__(cls, fmt, num_attrs, byte_order)
 
-    def __init__(self, fmt: str, num_attrs: int = 1, byte_order: ByteOrder = ByteOrder.DEFAULT, actions: tuple[Callable[[Any], Any], ...] = ()) -> None:
+    def __init__(
+        self,
+        fmt: str,
+        num_attrs: int = 1,
+        byte_order: ByteOrder = ByteOrder.DEFAULT,
+        actions: tuple[Callable[[Any], Any], ...] = (),
+    ) -> None:
         super().__init__(fmt, num_attrs, byte_order)
         if len(actions) < num_attrs:
             actions = tuple(
                 chain(actions, repeat(noop_action, num_attrs - len(actions)))
             )
         self.actions = actions
-
 
     def unpack(self, buffer: ReadableBuffer) -> tuple[Unpack[Ts]]:
         return tuple(
