+++ structured/utils.py
@@ -6,7 +6,7 @@ import sys
 import warnings
 from functools import wraps
 
-from .type_checking import Any, Callable, NoReturn, Optional, ParamSpec, T, TypeVar
+from .type_checking import Any, Callable, Optional, ParamSpec, T, TypeVar
 
 if sys.version_info < (3, 10):
     from typing import overload
@@ -64,32 +64,6 @@ def attrgetter(*attr_names: str) -> Callable[[Any], tuple[Any, ...]]:
         return lambda x: (_get(x),)
     else:
         return _get
-
-
-@classmethod
-def __error_getitem__(cls: type, _key: Any) -> NoReturn:
-    """Cause a helpful error if a class is indexed.  Used by `specialized`."""
-    raise TypeError(f'{cls.__qualname__} is already specialized.')
-
-
-def specialized(base_cls: type, *args: Any) -> Callable[[type[T]], type[T]]:
-    """Marks a class as already specialized, overriding the class' indexing
-    method with one that raises a helpful error.  Also fixes up the class'
-    qualname to be a more readable name.
-
-    :param cls: The class to mark as already specialized.
-    :return: The class with described modifications.
-    """
-
-    def wrapper(cls: type[T]) -> type[T]:
-        setattr(cls, '__class_getitem__', __error_getitem__)
-        qualname = ', '.join((getattr(k, '__qualname__', f'{k}') for k in args))
-        name = ', '.join((getattr(k, '__name__', f'{k}') for k in args))
-        cls.__qualname__ = f'{base_cls.__qualname__}[{qualname}]'
-        cls.__name__ = f'{base_cls.__name__}[{name}]'
-        return cls
-
-    return wrapper
 
 
 class StructuredAlias:
